refactor: split monolithic cstylecheck.py into package (issue #65)
Converts src/cstylecheck.py (4,077 lines) into src/cstylecheck/ package
with 9 focused sub-modules:
  models.py       — Violation, CheckResult, _ParamSig, _FuncSig
  preprocessor.py — source parsing utilities
  utils.py        — naming/case helpers
  config.py       — config/file loading
  checker.py      — Checker class and all _check_* methods
  sign_checker.py — SignChecker, DeclaredNotDefinedChecker
  baseline.py     — baseline suppression
  output.py       — Tee, JSON/SARIF formatters, summary
  cli.py          — CLI argument parsing and main()

__init__.py re-exports the full public API for backward compatibility.
src/cstylecheck.py is retained as a thin CLI wrapper so that the tool
can still be invoked directly as a script (python src/cstylecheck.py)
without installation, keeping all subprocess-based CLI tests working.
pyproject.toml updated to use package discovery.

Also updates CSC-SWE3-001 (docs/aspice/CStyleCheck_SWE3_Detailed_Design.md)
to v1.2: adds all 25+ missing units, correct line numbers, and documents
the target package structure.

All 812 tests pass.
